libs/ninja-gql/pyproject.toml
[project]
name = "ninja-gql"
version = "0.1.0"
description = "GraphQL layer generator for Ninja Stack (Strawberry)"
requires-python = ">=3.12"
dependencies = ["pydantic>=2.0", "strawberry-graphql>=0.236", "ninja-core", "ninja-persistence"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ninja_gql"]

[tool.uv.sources]
ninja-core = { workspace = true }
ninja-persistence = { workspace = true }
